change of url progress response: transfer values in t_uint instead of t_float removed gtkmm < 2.6 support added semtimedop function detection (FreeBSD6 hasn't it) using common humanReadable* stuff in url loading progress dialog,  removed unneded (duplicate) function Util::valueUnit

+++ trunk/daemon/http/curlif.cpp
@@ -159,7 +159,7 @@ namespace btg
          
          size_t curlInterface::curl_write_function(void *data, size_t size, size_t nmemb, curlInterface *self)
          {
-            return self->onDataChunk(data, size * nmemb) ? size * nmemb : -1;
+            return self->onDataChunk(data, size * nmemb) ? size * nmemb : (size_t)-1;
          }
          
          bool curlInterface::onDataChunk(void *data, size_t size)
